% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signs_concordances.R
\name{concordances}
\alias{concordances}
\title{Compute a vector of concordances from a collection of vector inputs}
\usage{
concordances(...)
}
\arguments{
\item{...}{≥2 vectors of input scores. These should all have the same length.
(Missing values?)}
}
\value{
A character vector summarizing the "concordances" of the inputs,
which are defined as the signs up to exchanging \code{-} and \code{+} and are denoted
using vertical bars and dashes (so that e.g. \verb{++-} and \verb{--+} yeild
concordance \verb{||-}) (Missing values?)
}
\description{
Compute a vector of concordances from a collection of vector inputs
}
