% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_gap.R
\name{rank_gap}
\alias{rank_gap}
\title{Compute rank-gap statistics from input scores}
\usage{
rank_gap(..., ties_method = "random", nonzero_adj = TRUE)
}
\arguments{
\item{...}{≥2 vectors of \emph{unsigned} input scores. These should all have
the same length.}

\item{ties_method}{Method of breaking ties; passed to \code{rank()}}

\item{nonzero_adj}{Whether to adjust ranks to avoid zero values of rank-gap
statistics}
}
\value{
A vector of rank-gap statistics, the same length as the input scores
}
\description{
Compute rank-gap statistics from input scores
}
