% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_gap.R
\name{rank_gap_df}
\alias{rank_gap_df}
\title{Conduct a rank-gap analysis; returns a data frame}
\usage{
rank_gap_df(
  ...,
  n_max_rank_bins = 1,
  ties_method = "random",
  nonzero_adj = TRUE
)
}
\description{
Conduct a rank-gap analysis; returns a data frame
}
