% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_gap.R
\name{rank_gap_hist}
\alias{rank_gap_hist}
\title{Produce a rank-gap histogram, with fill mapped to combinations of signs,
from signed input scores}
\usage{
rank_gap_hist(..., n_bins = 30, ties_method = "random", nonzero_adj = TRUE)
}
\arguments{
\item{...}{Signed input scores}

\item{n_bins}{Number of histogram bins}

\item{ties_method, nonzero_adj}{Arguments passed to \code{rank_gap()}}
}
\description{
Produce a rank-gap histogram, with fill mapped to combinations of signs,
from signed input scores
}
