% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_gap.R
\name{rank_gap_linehist}
\alias{rank_gap_linehist}
\title{Produce a rank-gap "line-histogram" based on signed input scores, with fill
mapped to combinations of signs}
\usage{
rank_gap_linehist(
  ...,
  n_max_rank_bins = 1,
  n_bins = 30,
  ties_method = "random",
  nonzero_adj = TRUE
)
}
\arguments{
\item{...}{Vectors of \emph{signed} input scores; these should be of equal
length}

\item{n_max_rank_bins}{Number of quantile bins for the maximum rank, to
focus on items ranked high across all input lists}

\item{n_bins}{Number of histogram bins}

\item{ties_method, nonzero_adj}{Arguments passed to \code{rank_gap()}}
}
\description{
Produce a rank-gap "line-histogram" based on signed input scores, with fill
mapped to combinations of signs
}
