% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rank_gap_qq}
\alias{rank_gap_qq}
\title{Produce a QQ plot of rank-gap statistics with color mapped to combinations
of signs.}
\usage{
rank_gap_qq(
  ...,
  n_max_rank_bins = 1,
  ties_method = "random",
  nonzero_adj = TRUE
)
}
\arguments{
\item{...}{Vectors of \emph{signed} input scores; these should be of equal
length}

\item{n_max_rank_bins}{Number of quantile bins for the maximum rank, to
focus on items ranked high in at least one list}

\item{ties_method, nonzero_adj}{Arguments passed to \code{rank_gap()}}
}
\description{
Produce a QQ plot of rank-gap statistics with color mapped to combinations
of signs.
}
