% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_gap.R
\name{signs_pm}
\alias{signs_pm}
\title{Compute a vector of signs from a collection of vector inputs}
\usage{
signs_pm(...)
}
\arguments{
\item{...}{≥2 vectors of numeric input scores. These should all have the same
length. (Missing values?)}
}
\value{
A character vector summarizing the signs of the inputs (Missing
values?)
}
\description{
Compute a vector of signs from a collection of vector inputs
}
