% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_gap.R
\name{rank_gap_df}
\alias{rank_gap_df}
\title{Conduct a rank-gap analysis; returns a data frame}
\usage{
rank_gap_df(
  ...,
  n_max_rank_bins = 1,
  ties_method = "random",
  nonzero_adj = TRUE
)
}
\arguments{
\item{...}{Vectors of \emph{signed} input scores; these should be of equal
length}

\item{n_max_rank_bins}{Number of quantile bins for the maximum rank, to
focus on items ranked high across all input lists}

\item{ties_method, nonzero_adj}{Arguments passed to \code{rank_gap()}}
}
\description{
Conduct a rank-gap analysis; returns a data frame
}
